Improve support for AFIO 64-bit inode values
>= 2^63 to zero values.  This turns out to be problematic for
filesystems that use very large inode values; it results
in all such files being viewed as hardlinks to each other.

values >= 2^63 as invalid and just ignoring them.  This prevented
the accidental hardlinking, but at the cost of losing all
hardlinks that involved large inode values.

This PR further improves things by stripping the high order bit
from 64-bit inode values in the AFIO parser.  This allows them
to be mostly preserved and should allow hardlinks to get properly
processed in the vast majority of cases.  The only false hardlinks
would be when there are inode values that differ exactly in the high
order bit, which should be very rare.

A full solution will require expanding inode handling to
use unsigned 64-bit values; we can't do that without a major
version bump, but this PR also sets the stage for migrating inode
support in a future libarchive 4.0.