zstd: Pad with a skippable frame of zeroes
Keep track of the total amount of data we've written, and on close, pad
with a skippable frame of zeroes.  If we don't, libarchive will pad with
just zeroes, which will result in an error when decompressing.

Although libzstd has an API for creating skippable frames, it is clunky
and marked experimental, so we do it ourselves.

Note that we could write out the frame header and let libarchive write
the zeroes, but this might break if there is another filter after ours.